Removed transformer function for OsmToShapely because it's not useful anymore.
To elaborate why it's not useful, generally when you want to transform the coordinates, it's to reproject the geometry to a new projection system. Before reprojection you should apply a mask to avoid out-of-bounds coordinates from being projected. OsmToShapely's transformer function may help in some situations, but most of the time it is not.